+++ core/CHANGELOG.md
@@ -1,6 +1,7 @@
 # TBD
 * Small comment/doc changes from weekly review
 * Easy fix to sort `api_container_env_vars` alphabetically
+* Remove some now-unneeded TODOs
 
 # 1.0.0
 * Add a tutorial explaining what Kurtosis does at the Docker level