+++ core/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Small comment/doc changes from weekly review
 * Easy fix to sort `api_container_env_vars` alphabetically
 * Remove some now-unneeded TODOs
+* Fix the `build_and_run.sh` script to use the proper way to pass in Docker args
 * Disallow test names with our test name delimiter: `,`
 
 # 1.0.0