+++ core/CHANGELOG.md
@@ -1,6 +1,9 @@
 # TBD
 * Small comment/doc changes from weekly review
 * Easy fix to sort `api_container_env_vars` alphabetically
+* Remove some now-unneeded TODOs
+* Fix the `build_and_run.sh` script to use the proper way to pass in Docker args
+* Disallow test names with our test name delimiter: `,`
 * Create an access controller with basic license auth
 * Connect access controller to auth0 device authorization flow
 * Implement machine-to-machine authorization flow for CI jobs